make-default-slots creates keywords
MAKE-DEFAULT-SLOTS used to make classes with symbols, but I changed it
so that it can create keywords using MAKE-KEYWORD.

Apparently the way MAKE-KEYWORD works is that it needs strings in all
caps otherwise it will make a weird looking keyword. So if I pass it
the string "abc" it will return keyword :|abc|, but if I pass "ABC" it
will return :ABC.

When creating the class I can use :abc or :ABC to initialize that slot
since both of those keywords are equal.

(eq :abc :ABC) ; => t